CHG: collection_public page to work with new public type [branches/20200806_acota_featured_collection_overhaul_q10271][q10271]
git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20200806_acota_featured_collection_overhaul_q10271@16110 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/collection_edit.php
@@ -113,7 +113,7 @@ if (getval("submitted","")!="" && enforcePostRequest(false))
 	}
 
 // Legacy property which is now superseeded by types. FCs need to be public before they can be put under a category by an admin (perm h)
-$collection["public"] = (in_array($collection["type"], array(COLLECTION_TYPE_PUBLIC, COLLECTION_TYPE_FEATURED)));
+$collection["public"] = (int) in_array($collection["type"], array(COLLECTION_TYPE_PUBLIC, COLLECTION_TYPE_FEATURED));
 
 include "../include/header.php";
 ?>
